backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    """Application Configuration"""
    
    # API Settings
    API_TITLE = "AI Secure Data Intelligence Platform"
    API_VERSION = "0.1.0"
    DEBUG = os.getenv("DEBUG", "False").lower() == "true"
    
    # File Processing
    MAX_FILE_SIZE_MB = 100
    MAX_FILE_SIZE_BYTES = MAX_FILE_SIZE_MB * 1024 * 1024
    ALLOWED_EXTENSIONS = {"txt", "log", "doc", "docx"}
    
    # AI Model Configuration - Groq LLM
    GROQ_API_KEY = os.getenv("GROQ_API_KEY", "")
    GROQ_MODEL = os.getenv("GROQ_MODEL", "llama-3.3-70b-versatile")
    
    # CORS - Allow Streamlit Cloud and local dev
    CORS_ORIGINS = ["*"]
    
    # Logging
    LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")

config = Config()
