requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
python-docx==0.8.11
groq==0.4.2
openai>=1.0.0
regex==2023.12.25
